requirements.txt
# Core Python utilities
python-dotenv>=1.0.0

# Web framework & server
fastapi>=0.110.0
uvicorn[standard]>=0.25.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Data validation & schemas
pydantic>=2.0.0

# HTTP requests (APIs, services, embeddings later)
requests>=2.31.0

# Reliability / retries
tenacity>=8.2.3

# LLM & embeddings
openai>=2.0.0
sentence-transformers>=2.6.0
torch>=2.0.0
